refactor(pds): move repo storage under repo/, split person-account tables into AccountStore
